+++ cdp-agentkit-core/typescript/src/actions/cdp/index.ts
@@ -11,8 +11,10 @@ import { TradeAction } from "./trade";
 import { TransferAction } from "./transfer";
 import { TransferNftAction } from "./transfer_nft";
 import { WrapEthAction } from "./wrap_eth";
-import { WOW_ACTIONS } from "./defi/wow";
+
+import { MORPHO_ACTIONS } from "./defi/morpho";
 import { PYTH_ACTIONS } from "./data/pyth";
+import { WOW_ACTIONS } from "./defi/wow";
 
 /**
  * Retrieves all CDP action instances.
@@ -37,7 +39,10 @@ export function getAllCdpActions(): CdpAction<CdpActionSchemaAny>[] {
   ];
 }
 
-export const CDP_ACTIONS = getAllCdpActions().concat(WOW_ACTIONS).concat(PYTH_ACTIONS);
+export const CDP_ACTIONS = getAllCdpActions()
+  .concat(MORPHO_ACTIONS)
+  .concat(PYTH_ACTIONS)
+  .concat(WOW_ACTIONS);
 
 export {
   CdpAction,